feat: Implement a remove command for server configurations
I think I'm going to stick with adding and removing configs instead
of editing details in-place. It'll reduce complexity, and adding a
new server config isn't too lengthy of a process. This commit will
confirm a removal, select a new current server if the removed one
was the current server, and save the new configuration.