feat: Implement LDIF output for search results
Sometimes saving search outputs to a file can be helpful. For
starters, this commit implements outputting LDIF without the extra
formatting and text of the 'standard' output. Saving in this format
is useful for backing up an entry, say before modifying it. This
commit leaves a placeholder for another useful format, CSV, but
does not implement it.