feat: Add command to modify attributes for an entry
This commit implements the command to add/edit/delete attributes
and/or their values. This is useful for small adjustments to existing
accounts that may be a lot of work to add to the account management
interface. Adding or removing entire entries/accounts is not an
intended feature of this command.